+++ src/library/output.rs
@@ -11,7 +11,7 @@ impl Node for Output {
     fn generate(
         &self,
         _position: &Coordinate,
-        _size: &(usize, usize),
+        _size: &(u32, u32),
         _input: InputOutputValue,
     ) -> Result<InputOutputValue> {
         todo!()
