+++ src/library/pattern.rs
@@ -12,7 +12,7 @@ impl Node for Pattern {
     fn generate(
         &self,
         position: &Coordinate,
-        _size: &(usize, usize),
+        _size: &(u32, u32),
         input: InputOutputValue,
     ) -> Result<InputOutputValue> {
         let p = input.to_common_ground()?;
