+++ src/library/static_value.rs
@@ -19,7 +19,7 @@ impl Node for StaticValue {
     fn generate(
         &self,
         _position: &Coordinate,
-        _size: &(usize, usize),
+        _size: &(u32, u32),
         _input: InputOutputValue,
     ) -> Result<InputOutputValue> {
         Ok(self.value)
